requirements.txt
asgiref==3.3.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.7.0rc2
Django==3.1.5
django-appconf==1.0.4
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-markdownfield==0.8.0
django-mdeditor==0.1.18
django-phonenumber-field==5.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.0.0
Markdown==3.3.3
oauthlib==3.1.0
packaging==20.7
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.16
pickleshare==0.7.5
pilkit==2.0
Pillow==8.0.1
pkg-resources==0.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
PyJWT==2.0.0
pyparsing==2.4.7
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
soupsieve==2.1
sqlparse==0.4.1
tqdm==4.56.0
traitlets==5.0.5
Unidecode==1.1.2
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
